エンティティを Y 座標順にソートし、その順番で描画するようにしました。
